Move function source code to subdirectory under src
