+++ src/Ahk.GitHub.Monitor/Config/CommentProtectionSettings.cs
