+++ src/Ahk.GitHub.Monitor/Config/EventSettingsBase.cs
