+++ src/Ahk.GitHub.Monitor/Config/MultiplePRProtectionSettings.cs
