+++ src/Ahk.GitHub.Monitor/Config/RepositorySettings.cs
