+++ src/Ahk.GitHub.Monitor/EventHandlers/EventHandlerResult.cs
