+++ src/Ahk.GitHub.Monitor/EventHandlers/IGitHubEventHandler.cs
