+++ src/Ahk.GitHub.Monitor/EventHandlers/IssueCommentEditDeleteHandler.cs
