+++ src/Ahk.GitHub.Monitor/Extensions/HeaderDictionaryExtensions.cs
