+++ src/Ahk.GitHub.Monitor/GitHubMonitorConfig.cs
