+++ src/Ahk.GitHub.Monitor/Helpers/PayloadReader.cs
