+++ src/Ahk.GitHub.Monitor/Helpers/PayloadValidator.cs
