+++ src/Ahk.GitHub.Monitor/Services/EventDispatchConfigBuilder.cs
