+++ src/Ahk.GitHub.Monitor/Services/EventDispatchService.cs
