+++ src/Ahk.GitHub.Monitor/Services/IEventDispatchService.cs
