+++ src/Ahk.GitHub.Monitor/host.json
